use skimmed ROOT file as cache (#63)
* use skim root file as cache

* missing

* up

* update readme

* fix test

* added missing

* install pyyaml

* add cachepath

* update singularity

* disable puid

* change to 4 cores

* tune memory/threads

* up

* remove obsolete code

* revert puid weight

* fix

* update hpc scripts

* ask for 4

* change file size

* fix

* fixes for condor

* fix

* printout

* some more docs

* bump ha version

* enable factorized jec

* comments

+++ batch/make_submit_jdl.sh
@@ -3,10 +3,15 @@ set -e
 
 #number of files to process per job
 #For factorized JEC, 5-50 is a good starting point
-export NCHUNK=20
-export SUBMIT_DIR=`pwd`/..
+export NCHUNK=10
+export SUBMIT_DIR=$(dirname `pwd`) 
 
 echo "Will create submit files based on directory SUBMIT_DIR="$SUBMIT_DIR
+
+if [[ ! -f ../out/datasets.json ]]; then
+    echo "../out/datasets.json does not exist, please create it using tests/hmm/run.sh or otherwise"
+    exit 1 
+fi
 
 #Clean old job files, copy from output directory
 rm -Rf jobfiles jobfiles.txt jobfiles.tgz
@@ -14,31 +19,33 @@ echo "Copying jobfiles"
 cp -R ../out/jobfiles ./
 
 #Create archive of job arguments
-echo "Creating archive"
+echo "Creating jobfile archive"
 tar -cvzf jobfiles.tgz jobfiles
 \ls -1 jobfiles/*.json | sed "s/jobfiles\///" | sed "s/\.json$//" > jobfiles.txt
 
-echo "Preparing job chunks"
 #Run N different random chunks per job
-python chunk_submits.py $NCHUNK > jobfiles_merged.txt
+echo "Preparing job chunks"
+python chunk_submits.py $NCHUNK jobfiles.txt > jobfiles_merged.txt
 
-#Prepare submit script
-cat analyze.jdl > submit.jdl
+rm -f args_analyze.txt
+rm -f slurm_submit.sh
 
 #Split on line, not on space
 IFS=$'\n'
 NJOB=0
+echo "Preparing condor argument file args_analyze.txt"
 for f in `cat jobfiles_merged.txt`; do
 
     #create condor submit files
-    echo "#NJOB="$NJOB >> submit.jdl
-    echo "Arguments = "$f >> submit.jdl
-    echo "Queue" >> submit.jdl
-    echo >> submit.jdl
+    echo $f >> args_analyze.txt 
 
-    #create SLURM submit file
+    #create SLURM submit file for HPC
     echo "sbatch slurm_hmm_analyze.sh "$f >> slurm_submit.sh
     NJOB=$((NJOB + 1))
 done
-echo "Please run 'export SUBMIT_DIR=`pwd`/..'"
+
+NJOBS=`wc -l args_analyze.txt`
+echo "Prepared jobs: $NJOBS"
+echo "Please run 'export SUBMIT_DIR=$SUBMIT_DIR'"
+echo "To submit the jobs, just run 'condor_submit analyze.jdl'"
